stt-provider/requirements.txt
openai-whisper
mysql-connector-python
fastapi
uvicorn
validators